Node API

ReactiveMP.Deterministic — Type

Deterministic

Deterministic object used to parametrize factor node object with determinstic type of relationship between variables.

See also: Stochastic, isdeterministic, isstochastic

ReactiveMP.Stochastic — Type

Stochastic

Stochastic object used to parametrize factor node object with stochastic type of relationship between variables.

See also: Deterministic, isdeterministic, isstochastic

ReactiveMP.isdeterministic — Function

isdeterministic(node)

Function used to check if factor node object is deterministic or not. Returns true or false.

See also: Deterministic, Stochastic, isstochastic

ReactiveMP.isstochastic — Function

isstochastic(node)

Function used to check if factor node object is stochastic or not. Returns true or false.

See also: Deterministic, Stochastic, isdeterministic

ReactiveMP.sdtype — Function

sdtype(object)

Returns either Deterministic or Stochastic for a given object (if defined).

See also: Deterministic, Stochastic, isdeterministic, isstochastic

ReactiveMP.MeanField — Type

MeanField

Generic factorisation constraint used to specify a mean-field factorisation for recognition distribution q.

See also: FullFactorisation

ReactiveMP.FullFactorisation — Type

FullFactorisation

Generic factorisation constraint used to specify a full factorisation for recognition distribution q.

See also: MeanField

ReactiveMP.collect_factorisation — Function

collect_factorisation(nodetype, factorisation)

This function converts given factorisation to a correct internal factorisation representation for a given node.

See also: MeanField, FullFactorisation

ReactiveMP.NodeInterface — Type

NodeInterface

NodeInterface object represents a single node-variable connection.

See also: name, tag, messageout, messagein

ReactiveMP.IndexedNodeInterface — Type

IndexedNodeInterface

IndexedNodeInterface object represents a repetative node-variable connection. Used in cases when node may connect different number of random variables with the same name, e.g. means and precisions of Gaussian Mixture node.

See also: name, tag, messageout, messagein

ReactiveMP.name — Function

name(interface)

Returns a name of the interface.

See also: NodeInterface, tag

ReactiveMP.tag — Function

tag(interface)

Returns a tag of the interface in the form of Val{ name(interface) }. The major difference between tag and name is that it is possible to dispath on interface's tag in message computation rule.

See also: NodeInterface, name

ReactiveMP.messageout — Function

messageout(interface)

Returns an outbound messages stream from the given interface.

See also: NodeInterface, messagein

ReactiveMP.messagein — Function

messagein(interface)

Returns an inbound messages stream from the given interface.

See also: NodeInterface, messageout

Internal API

ReactiveMP.connectvariable! — Function

connectvariable!(interface, variable, index)

Connects a variable with the interface and given index. Index is used to distinguish this connection from others in case if variable is connected to multiple interfaces.

See also: NodeInterface, connectedvar, connectedvarindex

ReactiveMP.connectedvar — Function

connectedvar(interface)

Returns connected variable for the interface.

See also: NodeInterface, connectvariable!, connectedvarindex

ReactiveMP.connectedvarindex — Function

connectedvarindex(interface)

Returns an index of connected variable for the interface.

See also: NodeInterface, connectvariable!, connectedvar

ReactiveMP.inbound_portal — Function

inbound_portal(interface)

Returns an instance of inbound portal of connected variable for the interface

See also: [`NodeInterface`](@ref), [`connectvariable!`](@ref), [`connectedvar`](@ref)

ReactiveMP.FactorNodeLocalMarginal — Type

FactorNodeLocalMarginal

This object represents local marginals for some specific factor node. Local marginal can be joint in case of structured factorisation. Local to factor node marginal also can be shared with a corresponding marginal of some random variable.

See also: FactorNodeLocalMarginals

ReactiveMP.FactorNodeLocalMarginals — Type

FactorNodeLocalMarginals

This object acts as an iterable and indexable proxy for local marginals for some node.