Distributions

ReactiveMP.jl library uses probability distributions types from Distributions.jl package. It also provides a wider range of possible parametrisations for some probability distributions, e.g NormalMeanPrecision or MvNormalMeanPrecision which might be more efficient in some situations.

API

ReactiveMP.vague — Function

vague(distribution_type, [ dims... ])

vague function returns uninformative probability distribution of a given type and can be used to set an uninformative priors in a model.

Base.prod — Function

prod(strategy, left, right)

prod function is used to find a product of two probability distrubution over same variable (e.g. 𝓝(x|μ1, σ1) × 𝓝(x|μ2, σ2)). There are two strategies for prod function: ProdPreserveParametrisation and ProdBestSuitableParametrisation.

Examples:

using ReactiveMP

product = prod(ProdPreserveParametrisation(), NormalMeanVariance(-1.0, 1.0), NormalMeanVariance(1.0, 1.0))

mean(product), var(product)

# output
(0.0, 0.5)

See also: default_prod_strategy, ProdPreserveParametrisation, ProdBestSuitableParametrisation

ReactiveMP.ProdPreserveParametrisation — Type

ProdPreserveParametrisation

ProdPreserveParametrisation is one of the strategies for prod function. This strategy assumes the same output factorisation (if possible). Should be used mostly to preserve distribution parametrisation across a model.

See also: prod, ProdBestSuitableParametrisation, ProdExpectationMaximisation

ReactiveMP.ProdBestSuitableParametrisation — Type

ProdBestSuitableParametrisation

ProdBestSuitableParametrisation is one of the strategies for prod function. This strategy does not make any assumptions about output factorisation. Can be used to speedup computations in some cases.

See also: prod, ProdPreserveParametrisation, ProdExpectationMaximisation

Missing docstring.

Missing docstring for default_prod_strategy. Check Documenter's build log for details.