Equality node

ReactiveMP.EqualityNode — Type

EqualityNode

Generic idea of an equality node is to keep track of intermediate left and right messages in a convenient manner so they can be reused during prod for random variables with large degree.

<- left –––- right -> ––- | = | ––– –––- | | outbound

Attributes:

left: An observable that indicates that left message can be computed, always sends missing since we are not interested in actual value. Can be treated as notification.

right: An observable that indicates that right message can be computed, always sends missing since we are not interested in actual value. Can be treated as notification.

cache_left: Keeps cached value for left outbound message, node itself does not track if cache is valid or not. Cache integrity is tracked by EqualityChain.

cache_right: Keeps cached value for right outbound message, node itself does not track if cache is valid or not. Cache integrity is tracked by EqualityChain.

See also: EqualityChain

ReactiveMP.EqualityChain — Type

EqualityChain