Helper utilities

OneDivNVector

Helper utilities implement OneDivNVector structure that is allocation free equivalent of fill(1 / N, N) collection. Mostly used in SampleList implementation.

ReactiveMP.OneDivNVector — Type

OneDivNVector(N::Int)
OneDivNVector(::Type{T}, N::Int) where T

Allocation-free version of fill(one(T) / N, N) vector.

Arguments

::Type{T}: type of elements, optional, Float64 by default, should be a subtype of Number

N::Int: number of elements in a container, should be greater than zero

Examples

julia> iter = ReactiveMP.OneDivNVector(3)
OneDivNVector(Float64, 3)

julia> length(iter)
3

julia> eltype(iter)
Float64

julia> collect(iter)
3-element Vector{Float64}:
 0.3333333333333333
 0.3333333333333333
 0.3333333333333333

julia> iter = ReactiveMP.OneDivNVector(Float32, 3)
OneDivNVector(Float32, 3)

julia> collect(iter)
3-element Vector{Float32}:
 0.33333334
 0.33333334
 0.33333334

See also: SampleList