Helper utilities

ReactiveMP implements various structures/functions/methods as "helper" structures that might be useful in various contexts.

SkipIndexIterator

ReactiveMP.SkipIndexIterator — Type

SkipIndexIterator

A special type of iterator that simply iterates over internal iterator, but skips index skip.

Arguments

iterator: internal iterator

skip: index to skip (integer)

See also: skipindex

ReactiveMP.skipindex — Function

skipindex(iterator, skip)

Creation operator for SkipIndexIterator.

julia> s = ReactiveMP.skipindex(1:3, 2)
2-element ReactiveMP.SkipIndexIterator{Int64, UnitRange{Int64}}:
 1
 3

julia> collect(s)
2-element Vector{Int64}:
 1
 3

See also: SkipIndexIterator

FunctionalIndex

ReactiveMP.FunctionalIndex — Type

FunctionalIndex

A special type of an index that represents a function that can be used only in pair with a collection. An example of a FunctionalIndex can be firstindex or lastindex, but more complex use cases are possible too, e.g. firstindex + 1. Important part of the implementation is that the resulting structure is isbitstype(...) = true, that allows to store it in parametric type as valtype.

One use case for this structure is to dispatch on and to replace begin or end (or more complex use cases, e.g. begin + 1) markers in constraints specification language.