Helper utilities

ReactiveMP implements various structures/function/method as "helper" structures that might be usefull in various contexts.

SkipIndexIterator

ReactiveMP.SkipIndexIterator — Type

SkipIndexIterator

A special type of iterator that simply iterates over internal iterator, but skips index skip.

Arguments

iterator: internal iterator

skip: index to skip (integer)

See also: skipindex

ReactiveMP.skipindex — Function

skipindex(iterator, skip)

Creation operator for SkipIndexIterator.

julia> s = ReactiveMP.skipindex(1:3, 2)
2-element ReactiveMP.SkipIndexIterator{Int64, UnitRange{Int64}}:
 1
 3

julia> collect(s)
2-element Vector{Int64}:
 1
 3

See also: SkipIndexIterator

InfCountingReal

ReactiveMP.InfCountingReal — Type

InfCountingReal

InfCountingReal implements a "number" that counts infinities in a separate field. Used to cancel out infinities in BFE computations. Infinity implemented as InfCountingReal(zero(Float64), 1).

Arguments

value::T: value of type <: Real

infs::Int: number of added/subtracted infinities

julia> r = ReactiveMP.InfCountingReal(0.0, 0)
InfCountingReal(0.0, 0∞)

julia> float(r)
0.0

julia> r = r + ReactiveMP.∞
InfCountingReal(0.0, 1∞)

julia> float(r)
Inf

julia> r = r - ReactiveMP.∞
InfCountingReal(0.0, 0∞)

julia> float(r)
0.0

See also: ∞

ReactiveMP.∞ — Constant

∞

A singleton object that implements "infinity" that can be added or subtracted to/from InfCountingReal structure. Internally implemented as InfCountingReal(zero(Float64), 0).

See also: InfCountingReal

deep_eltype

ReactiveMP.deep_eltype — Function

deep_eltype

Returns the eltype of the first container in the nested hierarchy.

julia> ReactiveMP.deep_eltype([ [1, 2], [2, 3] ])
Int64

julia> ReactiveMP.deep_eltype([[[ 1.0, 2.0 ], [ 3.0, 4.0 ]], [[ 5.0, 6.0 ], [ 7.0, 8.0 ]]])
Float64

FunctionalIndex

ReactiveMP.FunctionalIndex — Type

FunctionalIndex

A special type of an index that represents a function that can be used only in pair with a collection. An example of a FunctionalIndex can be firstindex or lastindex, but more complex use cases are possible too, e.g. firstindex + 1. Important part of the implementation is that the resulting structure is isbitstype(...) = true, that allows to store it in parametric type as valtype.

One use case for this structure is to dispatch on and to replace begin or end (or more complex use cases, e.g. begin + 1) markers in constraints specification language.