BinomialPolya node

The BinomialPolya node implements a Binomial likelihood with logistic linear predictor and PolyaGamma augmentation for Bayesian inference. This node is particularly useful for modeling count data with overdispersion and performing Binomial regression.

ReactiveMP.BinomialPolya — Type

BinomialPolya

A node type representing a Binomial likelihood with linear predictor through logistic. A Normal prior on the weights is used. The prior is augmented with a PolyaGamma distribution, which is used for modeling count data with overdispersion. This implementation follows the PolyaGamma augmentation scheme for Bayesian inference. Can be used for Binomial regression.

ReactiveMP.BinomialPolyaMeta — Type

BinomialPolyaMeta

Metadata structure for the BinomialPolya node. It will be passed to rules. In case no meta is provided, the rules will use the means to compute the messages. Both schemes yield very similar results.

Fields

n_samples::Int: Number of samples to use for Monte Carlo estimation of the average energy. Default is 1, as increasing it adds computational cost without significant benefit.

rng::AbstractRNG: Random number generator to use for sampling. Defaults to Random.default_rng().