Rules implementation

Message update rules

ReactiveMP.rule — Function

rule(fform, on, vconstraint, mnames, messages, qnames, marginals, meta, __node)

This function is used to compute an outbound message for a given node

Arguments

fform: Functional form of the node in form of a type of the node, e.g. ::Type{ <: NormalMeanVariance } or ::typeof(+)

on: Outbound interface's tag for which a message has to be computed, e.g. ::Val{:out} or ::Val{:μ}

vconstraint: Variable constraints for an outbound interface, e.g. Marginalisation or MomentMatching

mnames: Ordered messages names in form of the Val type, eg. ::Val{ (:mean, :precision) }

messages: Tuple of message of the same length as mnames used to compute an outbound message

qnames: Ordered marginal names in form of the Val type, eg. ::Val{ (:mean, :precision) }

marginals: Tuple of marginals of the same length as qnames used to compute an outbound message

meta: Extra meta information

addons: Extra addons information

__node: Node reference

For all available rules, see ReactiveMP.print_rules_table().

See also: @rule, marginalrule, @marginalrule

ReactiveMP.@rule — Macro

@rule NodeType(:Edge, Constraint) (Arguments..., [ meta::MetaType ]) = begin
    # rule body
    return ...
end

The @rule macro help to define new methods for the rule function. It works particularly well in combination with the @node macro. It has a specific structure, which must specify:

NodeType: must be a valid Julia type. If some attempt to define a rule for a Julia function (for example +), use typeof(+)

Edge: edge label, usually edge labels are defined with the @node macro

Constrain: DEPRECATED, please just use the Marginalisation label

Arguments: defines a list of the input arguments for the rule

m_* prefix indicates that the argument is of type Message from the edge *

q_* prefix indicates that the argument is of type Marginal on the edge *

Meta::MetaType - optionally, a user can specify a Meta object of type MetaType. This can be useful is some attempts to try different rules with different approximation methods or if the rule itself requires some temporary storage or cache. The default meta is nothing.

Here are various examples of the @rule macro usage:

Belief-Propagation (or Sum-Product) message update rule for the NormalMeanVariance node toward the :μ edge with the Marginalisation constraint. Input arguments are m_out and m_v, which are the messages from the corresponding edges out and v and have the type PointMass.

@rule NormalMeanVariance(:μ, Marginalisation) (m_out::PointMass, m_v::PointMass) = NormalMeanVariance(mean(m_out), mean(m_v))

Mean-field message update rule for the NormalMeanVariance node towards the :μ edge with the Marginalisation constraint. Input arguments are q_out and q_v, which are the marginals on the corresponding edges out and v of type Any.

@rule NormalMeanVariance(:μ, Marginalisation) (q_out::Any, q_v::Any) = NormalMeanVariance(mean(q_out), mean(q_v))

Structured Variational message update rule for the NormalMeanVariance node towards the :out edge with the Marginalisation constraint. Input arguments are m_μ, which is a message from the μ edge of type UnivariateNormalDistributionsFamily, and q_v, which is a marginal on the v edge of type Any.

@rule NormalMeanVariance(:out, Marginalisation) (m_μ::UnivariateNormalDistributionsFamily, q_v::Any) = begin
    m_μ_mean, m_μ_cov = mean_cov(m_μ)
    return NormalMeanVariance(m_μ_mean, m_μ_cov + mean(q_v))
end

See also: rule, marginalrule, [@marginalrule], @call_rule

ReactiveMP.@call_rule — Macro

@call_rule NodeType(:edge, Constraint) (argument1 = value1, argument2 = value2, ..., [ meta = ..., addons = ... ])

The @call_rule macro helps to call the rule method with an easier syntax. The structure of the macro is almost the same as in the @rule macro, but there is no begin ... end block, but instead each argument must have a specified value with the = operator.

The @call_rule accepts optional list of options before the functional form specification, for example:

@call_rule [ return_addons = true ] NodeType(:edge, Constraint) (argument1 = value1, argument2 = value2, ..., [ meta = ..., addons = ... ])

The list of available options is:

return_addons - forces the @call_rule to return the tuple of (result, addons)

See also: @rule, rule, @call_marginalrule

Marginal update rules

ReactiveMP.marginalrule — Function

marginalrule(fform, on, mnames, messages, qnames, marginals, meta, __node)

This function is used to compute a local joint marginal for a given node

Arguments

fform: Functional form of the node in form of a type of the node, e.g. ::Type{ <: NormalMeanVariance } or ::typeof(+)

on: Local joint marginal tag , e.g. ::Val{ :mean_precision } or ::Val{ :out_mean_precision }

mnames: Ordered messages names in form of the Val type, eg. ::Val{ (:mean, :precision) }

messages: Tuple of message of the same length as mnames used to compute an outbound message

qnames: Ordered marginal names in form of the Val type, eg. ::Val{ (:mean, :precision) }

marginals: Tuple of marginals of the same length as qnames used to compute an outbound message

meta: Extra meta information

__node: Node reference

See also: rule, @rule @marginalrule

ReactiveMP.@marginalrule — Macro

@marginalrule NodeType(:Cluster) (Arguments..., [ meta::MetaType ]) = begin
    # rule body
    return ...
end

The @marginalrule macro help to define new methods for the marginalrule function. It works particularly well in combination with the @node macro. It has a specific structure, which must specify:

NodeType: must be a valid Julia type. If some attempt to define a rule for a Julia function (for example +), use typeof(+)

Cluster: edge cluster that contains joined edge labels with the _ symbol. Usually edge labels are defined with the @node macro

Arguments: defines a list of the input arguments for the rule

m_* prefix indicates that the argument is of type Message from the edge *

q_* prefix indicates that the argument is of type Marginal on the edge *

Meta::MetaType - optionally, a user can specify a Meta object of type MetaType. This can be useful is some attempts to try different rules with different approximation methods or if the rule itself requires some temporary storage or cache. The default meta is nothing.

The @marginalrule can return a NamedTuple in the return statement. This would indicate some variables in the joint marginal for the Cluster are independent and the joint itself is factorised. For example if some attempts to compute a marginal for the q(x, y) it is possible to return (x = ..., y = ...) as the result of the computation to indicate that q(x, y) = q(x)q(y).

Here are various examples of the @marginalrule macro usage:

Marginal computation rule around the NormalMeanPrecision node for the q(out, μ). The rule accepts arguments m_out and m_μ, which are the messages

from the out and μ edges respectively, and q_τ which is the marginal on the edge τ.

@marginalrule NormalMeanPrecision(:out_μ) (m_out::UnivariateNormalDistributionsFamily, m_μ::UnivariateNormalDistributionsFamily, q_τ::Any) = begin
    xi_out, W_out = weightedmean_precision(m_out)
    xi_μ, W_μ     = weightedmean_precision(m_μ)

    W_bar = mean(q_τ)

    W  = [W_out+W_bar -W_bar; -W_bar W_μ+W_bar]
    xi = [xi_out; xi_μ]

    return MvNormalWeightedMeanPrecision(xi, W)
end

Marginal computation rule around the NormalMeanPrecision node for the q(out, μ). The rule accepts arguments m_out and m_μ, which are the messages from the

out and μ edges respectively, and q_τ which is the marginal on the edge τ. In this example the result of the computation is a NamedTuple

@marginalrule NormalMeanPrecision(:out_μ) (m_out::PointMass, m_μ::UnivariateNormalDistributionsFamily, q_τ::Any) = begin
    return (out = m_out, μ = prod(ProdAnalytical(), NormalMeanPrecision(mean(m_out), mean(q_τ)), m_μ))
end

ReactiveMP.@call_marginalrule — Macro

@call_marginalrule NodeType(:edge) (argument1 = value1, argument2 = value2, ..., [ meta = ... ])

The @call_marginalrule macro helps to call the marginalrule method with an easier syntax. The structure of the macro is almost the same as in the @marginalrule macro, but there is no begin ... end block, but instead each argument must have a specified value with the = operator.

See also: @marginalrule, marginalrule, @call_rule

Testing utilities for the update rules

ReactiveMP.@test_rules — Macro

@test_rules [options] rule [ test_entries... ]

The @test_rules macro generates test cases for message update rules for probabilistic programming models that follow the "message passing" paradigm. It takes a rule specification as input and generates a set of tests based on that specification. This macro is provided by ReactiveMP.

Note: The Test module must be imported explicitly. The @test_rules macro tries to use the @test macro, which must be defined globally.

Arguments

The macro takes three arguments:

options: An optional argument that specifies the options for the test generation process. See below for details.

rule: A rule specification in the same format as the @rule macro, e.g. Beta(:out, Marginalisation) or NormalMeanVariance(:μ, Marginalisation).

test_entries: An array of named tuples (input = ..., output = ...). The input entry has the same format as the input for the @rule macro. The output entry specifies the expected output.

Options

The following options are available:

check_type_promotion: By default, this option is set to false. If set to true, the macro generates an extensive list of extra tests that aim to check the correct type promotion within the tests. For example, if all inputs are of type Float32, then the expected output should also be of type Float32. See the paramfloattype and convert_paramfloattype functions for details.

atol: Sets the desired accuracy for the tests. The tests use the custom_isapprox function from ReactiveMP to check if outputs are approximately the same. This argument can be either a single number or an array of key => value pairs.

extra_float_types: A set of extra float types to be used in the check_type_promotion tests. This argument has no effect if check_type_promotion is set to false.

The default values for the atol option are:

Float32: 1e-4

Float64: 1e-6

BigFloat: 1e-8

Examples

@test_rules [check_type_promotion = true] Beta(:out, Marginalisation) [
    (input = (m_a = PointMass(1.0), m_b = PointMass(2.0)), output = Beta(1.0, 2.0)),
    (input = (m_a = PointMass(2.0), m_b = PointMass(2.0)), output = Beta(2.0, 2.0)),
    (input = (m_a = PointMass(3.0), m_b = PointMass(3.0)), output = Beta(3.0, 3.0))
]

@test_rules [check_type_promotion = true] Beta(:out, Marginalisation) [
    (input = (q_a = PointMass(1.0), q_b = PointMass(2.0)), output = Beta(1.0, 2.0)),
    (input = (q_a = PointMass(2.0), q_b = PointMass(2.0)), output = Beta(2.0, 2.0)),
    (input = (q_a = PointMass(3.0), q_b = PointMass(3.0)), output = Beta(3.0, 3.0))
]

See also: ReactiveMP.@test_marginalrules

ReactiveMP.@test_marginalrules — Macro

@test_marginalrules [options] rule [ test_entries... ]

Effectively the same as @test_rules, but for marginal computation rules. See the documentation for @test_rules for more info.

See also: ReactiveMP.@test_rules

Table of available update rules

using ReactiveMP, Markdown
Markdown.parse(ReactiveMP.print_rules_table())

Node	Output	Inputs	Meta

AR	y	μ(x) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	ARMeta

q(θ) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

q(γ) :: Any

AR	y	q(x) :: Any	ARMeta

q(θ) :: Any

q(γ) :: Any

AR	x	μ(y) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	ARMeta

q(θ) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

q(γ) :: Any

AR	x	q(y) :: Any	ARMeta

q(θ) :: Any

q(γ) :: Any

AR	θ	q(y_x) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	ARMeta

q(γ) :: Any

AR	θ	q(y) :: Any	ARMeta

q(x) :: Any

q(γ) :: Any

AR	γ	q(y_x) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	ARMeta

q(θ) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

AR	γ	q(y) :: Any	ARMeta

q(x) :: Any

q(θ) :: Any

MvNormalMeanCovariance	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	out	μ(μ) :: PointMass	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	out	q(μ) :: PointMass	Nothing

q(Σ) :: PointMass

MvNormalMeanCovariance	out	q(μ) :: Any	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(Σ) :: PointMass

MvNormalMeanCovariance	out	μ(μ) :: PointMass	Nothing

μ(Σ) :: PointMass

MvNormalMeanCovariance	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	μ	μ(out) :: PointMass	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	μ	q(out) :: PointMass	Nothing

q(Σ) :: PointMass

MvNormalMeanCovariance	μ	q(out) :: Any	Nothing

q(Σ) :: Any

MvNormalMeanCovariance	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(Σ) :: PointMass

MvNormalMeanCovariance	μ	μ(out) :: PointMass	Nothing

μ(Σ) :: PointMass

MvNormalMeanCovariance	Σ	q(out) :: Any	Nothing

q(μ) :: Any

MvNormalMeanCovariance	Σ	q(out_μ) :: Any	Nothing

InverseWishart	out	μ(ν) :: PointMass	Nothing

μ(S) :: PointMass

InverseWishart	out	μ(S) :: PointMass	Nothing

q(ν) :: Any

InverseWishart	out	μ(ν) :: PointMass	Nothing

q(S) :: Any

InverseWishart	out	q(ν) :: Any	Nothing

q(S) :: Any

Dirichlet	out	μ(a) :: PointMass{<:AbstractVector}	Nothing

Dirichlet	out	q(a) :: PointMass{<:AbstractVector}	Nothing

GCV	x	μ(y) :: Union{ExponentialLinearQuadratic, UnivariateNormalDistributionsFamily, UnivariateGaussianDistributionsFamily}	Union{Nothing, GCVMetadata}

q(z) :: Any

q(κ) :: Any

q(ω) :: Any

GCV	x	q(y) :: Any	Union{Nothing, GCVMetadata}

q(z) :: Any

q(κ) :: Any

q(ω) :: Any

GCV	y	μ(x) :: Union{ExponentialLinearQuadratic, UnivariateNormalDistributionsFamily, UnivariateGaussianDistributionsFamily}	Union{Nothing, GCVMetadata}

q(z) :: Any

q(κ) :: Any

q(ω) :: Any

GCV	y	q(x) :: Any	Union{Nothing, GCVMetadata}

q(z) :: Any

q(κ) :: Any

q(ω) :: Any

GCV	z	q(y_x) :: Any	GCVMetadata

q(κ) :: Any

q(ω) :: Any

GCV	z	q(y) :: Any	GCVMetadata

q(x) :: Any

q(κ) :: Any

q(ω) :: Any

GCV	κ	q(y_x) :: Any	GCVMetadata

q(z) :: Any

q(ω) :: Any

GCV	κ	q(y) :: Any	GCVMetadata

q(x) :: Any

q(z) :: Any

q(ω) :: Any

GCV	ω	q(y_x) :: Any	GCVMetadata

q(z) :: Any

q(κ) :: Any

GCV	ω	q(y) :: Any	GCVMetadata

q(x) :: Any

q(z) :: Any

q(κ) :: Any

Distributions.InverseGamma	out	μ(α) :: PointMass	Nothing

μ(θ) :: PointMass

Distributions.InverseGamma	out	q(α) :: Any	Nothing

q(θ) :: Any

MatrixDirichlet	out	μ(a) :: PointMass	Nothing

MatrixDirichlet	out	q(a) :: PointMass	Nothing

MvNormalMeanPrecision	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Λ) :: Wishart

MvNormalMeanPrecision	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	out	μ(μ) :: PointMass	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	out	q(μ) :: PointMass	Nothing

q(Λ) :: PointMass

MvNormalMeanPrecision	out	q(μ) :: Any	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(Λ) :: PointMass

MvNormalMeanPrecision	out	μ(μ) :: PointMass	Nothing

μ(Λ) :: PointMass

MvNormalMeanPrecision	μ	μ(out) :: PointMass	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	μ	q(out) :: PointMass	Nothing

q(Λ) :: PointMass

MvNormalMeanPrecision	μ	q(out) :: Any	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(Λ) :: PointMass

MvNormalMeanPrecision	μ	μ(out) :: PointMass	Nothing

μ(Λ) :: PointMass

MvNormalMeanPrecision	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Λ) :: Wishart

MvNormalMeanPrecision	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(Λ) :: Any

MvNormalMeanPrecision	Λ	q(out) :: Any	Nothing

q(μ) :: Any

MvNormalMeanPrecision	Λ	q(out_μ) :: Any	Nothing

Gamma	out	μ(α) :: PointMass	Nothing

μ(θ) :: PointMass

Gamma	out	q(α) :: Any	Nothing

q(θ) :: Any

Poisson	l	μ(out) :: PointMass	Nothing

Poisson	l	q(out) :: Any	Nothing

Poisson	out	μ(l) :: PointMass	Nothing

Poisson	out	q(l) :: GammaDistributionsFamily	Nothing

Flow	in	μ(out) :: MvNormalMeanCovariance	FlowMeta{M, <:Linearization}

Flow	in	μ(out) :: MvNormalMeanPrecision	FlowMeta{M, <:Linearization}

Flow	in	μ(out) :: MvNormalWeightedMeanPrecision	FlowMeta{M, <:Linearization}

Flow	in	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	FlowMeta{M, <:Unscented}

Flow	out	μ(in) :: MvNormalMeanCovariance	FlowMeta{M, <:Linearization}

Flow	out	μ(in) :: MvNormalMeanPrecision	FlowMeta{M, <:Linearization}

Flow	out	μ(in) :: MvNormalWeightedMeanPrecision	FlowMeta{M, <:Linearization}

Flow	out	μ(in) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	FlowMeta{M, <:Unscented}

NormalMeanPrecision	out	μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

q(τ) :: Any

NormalMeanPrecision	out	μ(μ) :: PointMass	Nothing

q(τ) :: Any

NormalMeanPrecision	out	q(μ) :: PointMass	Nothing

q(τ) :: PointMass

NormalMeanPrecision	out	q(μ) :: Any	Nothing

q(τ) :: Any

NormalMeanPrecision	out	μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(τ) :: PointMass

NormalMeanPrecision	out	μ(μ) :: PointMass	Nothing

μ(τ) :: PointMass

NormalMeanPrecision	μ	μ(out) :: PointMass	Nothing

q(τ) :: Any

NormalMeanPrecision	μ	q(out) :: PointMass	Nothing

q(τ) :: PointMass

NormalMeanPrecision	μ	q(out) :: Any	Nothing

q(τ) :: Any

NormalMeanPrecision	μ	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(τ) :: PointMass

NormalMeanPrecision	μ	μ(out) :: PointMass	Nothing

μ(τ) :: PointMass

NormalMeanPrecision	μ	μ(out) :: ExponentialLinearQuadratic

q(τ) :: Any

NormalMeanPrecision	μ	μ(out) :: ExponentialLinearQuadratic

μ(τ) :: PointMass

NormalMeanPrecision	μ	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

q(τ) :: Any

NormalMeanPrecision	τ	q(out) :: Any	Nothing

q(μ) :: Any

NormalMeanPrecision	τ	q(out_μ) :: Any	Nothing

NormalMeanVariance	out	μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

q(v) :: PointMass

NormalMeanVariance	out	μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

q(v) :: Any

NormalMeanVariance	out	μ(μ) :: PointMass	Nothing

q(v) :: Any

NormalMeanVariance	out	q(μ) :: PointMass	Nothing

q(v) :: PointMass

NormalMeanVariance	out	q(μ) :: Any	Nothing

q(v) :: Any

NormalMeanVariance	out	μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(v) :: PointMass

NormalMeanVariance	out	μ(μ) :: PointMass	Nothing

μ(v) :: PointMass

NormalMeanVariance	μ	q(out) :: PointMass	Nothing

q(v) :: PointMass

NormalMeanVariance	μ	q(out) :: Any	Nothing

q(v) :: Any

NormalMeanVariance	μ	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(v) :: PointMass

NormalMeanVariance	μ	μ(out) :: PointMass	Nothing

μ(v) :: PointMass

NormalMeanVariance	μ	μ(out) :: ExponentialLinearQuadratic

q(v) :: Any

NormalMeanVariance	μ	μ(out) :: ExponentialLinearQuadratic

μ(v) :: PointMass

NormalMeanVariance	μ	μ(out) :: PointMass	Nothing

q(v) :: Any

NormalMeanVariance	μ	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

q(v) :: Any

NormalMeanVariance	v	μ(out) :: PointMass	Nothing

μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

NormalMeanVariance	v	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(μ) :: PointMass

NormalMeanVariance	v	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(μ) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

NormalMeanVariance	v	q(out) :: Any	Nothing

q(μ) :: Any

NormalMeanVariance	v	q(out_μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

Beta	out	μ(a) :: PointMass	Nothing

μ(b) :: PointMass

Beta	out	q(a) :: PointMass	Nothing

q(b) :: PointMass

Wishart	out	μ(ν) :: PointMass	Nothing

μ(S) :: PointMass

Wishart	out	μ(S) :: PointMass	Nothing

q(ν) :: Any

Wishart	out	μ(ν) :: PointMass	Nothing

q(S) :: Any

Wishart	out	q(ν) :: Any	Nothing

q(S) :: Any

Uniform	out	μ(a) :: PointMass	Nothing

μ(b) :: PointMass

Uniform	out	μ(b) :: PointMass	Nothing

q(a) :: PointMass

Uniform	out	μ(a) :: PointMass	Nothing

q(b) :: PointMass

Uniform	out	q(a) :: PointMass	Nothing

q(b) :: PointMass

Categorical{P} where P<:Real	out	μ(p) :: Dirichlet	Nothing

Categorical{P} where P<:Real	out	q(p) :: Dirichlet	Nothing

Categorical{P} where P<:Real	out	μ(p) :: PointMass	Nothing

Categorical{P} where P<:Real	out	q(p) :: PointMass	Nothing

Categorical{P} where P<:Real	p	q(out) :: PointMass	Nothing

Categorical{P} where P<:Real	p	q(out) :: Any	Nothing

NOT	in	μ(out) :: Bernoulli	Nothing

NOT	out	μ(in) :: Bernoulli	Nothing

Bernoulli	p	μ(out) :: PointMass	Nothing

Bernoulli	p	q(out) :: PointMass	Nothing

Bernoulli	p	q(out) :: Bernoulli	Nothing

Bernoulli	p	q(out) :: Categorical{P} where P<:Real	Nothing

Bernoulli	out	μ(p) :: Beta	Nothing

Bernoulli	out	μ(p) :: PointMass	Nothing

Bernoulli	out	q(p) :: PointMass	Nothing

Bernoulli	out	q(p) :: Any	Nothing

AND	in1	μ(out) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

AND	in2	μ(out) :: Bernoulli	Nothing

μ(in1) :: Bernoulli

AND	out	μ(in1) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

Transition	out	μ(in) :: Categorical{P} where P<:Real	Nothing

μ(a) :: PointMass

Transition	out	q(in) :: PointMass	Nothing

q(a) :: PointMass

Transition	out	q(in) :: Categorical{P} where P<:Real	Nothing

q(a) :: Any

Transition	out	μ(in) :: Categorical{P} where P<:Real	Nothing

q(a) :: Distributions.Distribution{Distributions.Matrixvariate, Distributions.Continuous}

Transition	out	μ(in) :: Distributions.DiscreteNonParametric

q(a) :: PointMass

Transition	in	μ(out) :: Categorical{P} where P<:Real	Nothing

μ(a) :: PointMass

Transition	in	q(out) :: Any	Nothing

q(a) :: MatrixDirichlet

Transition	in	μ(out) :: Categorical{P} where P<:Real	Nothing

q(a) :: MatrixDirichlet

Transition	in	μ(out) :: Categorical{P} where P<:Real

q(a) :: PointMass

Transition	in	q(out) :: PointMass	Nothing

q(a) :: PointMass

Transition	a	q(out) :: Any	Nothing

q(in) :: Categorical{P} where P<:Real

Transition	a	q(out_in) :: Contingency	Nothing

NormalMixture{N}	switch	q(out) :: Any	Nothing

q(m) :: ManyOf{<:Tuple{Vararg{Any, N}}}

q(p) :: ManyOf{<:Tuple{Vararg{Any, N}}}

NormalMixture	m	q(out) :: Any	Nothing

q(switch) :: Any

q(p) :: Any

NormalMixture	p	q(out) :: Any	Nothing

q(switch) :: Any

q(m) :: Any

NormalMixture{N}	out	q(switch) :: Any	Nothing

q(m) :: ManyOf{<:Tuple{Vararg{Any, N}}}

q(p) :: ManyOf{<:Tuple{Vararg{Any, N}}}

Probit	in	q(out) :: PointMass	Union{Nothing, ProbitMeta}

Probit	in	μ(out) :: Union{Bernoulli, PointMass}	Union{Nothing, ProbitMeta}

Probit	in	μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, ProbitMeta}

q(out) :: PointMass

Probit	in	μ(out) :: Union{Bernoulli, PointMass}	Union{Nothing, ProbitMeta}

μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

Probit	out	μ(in) :: PointMass	Union{Nothing, ProbitMeta}

Probit	out	q(in) :: PointMass	Union{Nothing, ProbitMeta}

Probit	out	μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, ProbitMeta}

Mixture	inputs	μ(out) :: Any	Nothing

μ(switch) :: Any

Mixture	inputs	μ(out) :: Any	Nothing

q(switch) :: PointMass

Mixture	out	μ(switch) :: Any	Nothing

μ(inputs) :: ManyOf{<:Tuple{Vararg{Any, N}}}

Mixture	out	μ(inputs) :: ManyOf{<:Tuple{Vararg{Any, N}}}	Nothing

q(switch) :: PointMass

Mixture	switch	μ(out) :: Any	Nothing

μ(inputs) :: ManyOf{<:Tuple{Vararg{Any, N}}}

GammaShapeRate	out	μ(α) :: PointMass	Nothing

μ(β) :: PointMass

GammaShapeRate	out	q(α) :: PointMass	Nothing

q(β) :: PointMass

GammaShapeRate	α	q(out) :: Gamma	Nothing

q(β) :: Gamma

GammaMixture{N}	switch	q(out) :: Any	Nothing

q(a) :: ManyOf{<:Tuple{Vararg{Any, N}}}

q(b) :: GammaDistributionsFamily

GammaMixture	a	q(out) :: Any	Nothing

q(switch) :: Any

q(b) :: GammaDistributionsFamily

GammaMixture	b	q(out) :: Any	Nothing

q(switch) :: Any

q(a) :: Any

GammaMixture{N}	out	q(switch) :: Any	Nothing

q(a) :: ManyOf{<:Tuple{Vararg{Any, N}}}

q(b) :: GammaDistributionsFamily

BIFMHelper	in	μ(out) :: Any	Nothing

BIFMHelper	out	q(in) :: Any	Nothing

BIFM	in	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	BIFMMeta

μ(zprev) :: ProdFinal{<:Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T}

μ(znext) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

BIFM	out	μ(in) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	BIFMMeta

μ(zprev) :: ProdFinal{<:Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T}

μ(znext) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

BIFM	znext	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	BIFMMeta

μ(in) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

μ(zprev) :: ProdFinal{<:Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T}

BIFM	zprev	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	BIFMMeta

μ(in) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

μ(znext) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

MvNormalMeanScalePrecision	out	q(μ) :: Any	Nothing

q(γ) :: Any

MvNormalMeanScalePrecision	out	μ(μ) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(γ) :: Any

MvNormalMeanScalePrecision	μ	q(out) :: Any	Nothing

q(γ) :: Any

MvNormalMeanScalePrecision	μ	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

q(γ) :: Any

MvNormalMeanScalePrecision	γ	q(out) :: Any	Nothing

q(μ) :: Any

MvNormalMeanScalePrecision	γ	q(out_μ) :: Any	Nothing

SoftDot	y	q(θ) :: Any	Nothing

q(x) :: Any

q(γ) :: Any

SoftDot	x	q(y) :: Any	Nothing

q(θ) :: Any

q(γ) :: Any

SoftDot	θ	q(y) :: Any	Nothing

q(x) :: Any

q(γ) :: Any

SoftDot	γ	q(y) :: Any	Nothing

q(θ) :: Any

q(x) :: Any

OR	in1	μ(out) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

in1	μ(out) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

OR	in2	μ(out) :: Bernoulli	Nothing

μ(in1) :: Bernoulli

in2	μ(out) :: Bernoulli	Nothing

μ(in1) :: Bernoulli

OR	out	μ(in1) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

out	μ(in1) :: Bernoulli	Nothing

μ(in2) :: Bernoulli

HalfNormal	out	q(v) :: PointMass	Nothing

DeltaFn	in	μ(in) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, Nothing}

q(ins) :: JointNormal

DeltaFn	in	μ(out) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, I}

μ(ins) :: Nothing

DeltaFn	in	μ(out) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, I}

μ(ins) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

DeltaFn	out	μ(ins) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M}

DeltaFn	in	μ(in) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, I}

q(ins) :: JointNormal

DeltaFn	in	μ(out) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, I}

μ(ins) :: Nothing

DeltaFn	in	μ(out) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M, I}

μ(ins) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

DeltaFn	out	μ(ins) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	DeltaMeta{M}

DeltaFn	in	μ(in) :: Any	DeltaMeta{M}

q(ins) :: ReactiveMP.FactorizedJoint

DeltaFn	out	q(ins) :: ReactiveMP.FactorizedJoint{P}	DeltaMeta{M}

DeltaFn	out	q(ins) :: ReactiveMP.FactorizedJoint	DeltaMeta{M}

-	out	μ(in1) :: Any

μ(in2) :: Any

-	in1	μ(out) :: Any

μ(in2) :: Any

-	in2	μ(out) :: Any

μ(in1) :: Any

dot	out	μ(in1) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in2) :: PointMass

dot	out	μ(in1) :: PointMass	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in2) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

dot	in1	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in2) :: PointMass

dot	in2	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in1) :: PointMass

+	out	μ(in1) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(in2) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: NormalMeanPrecision	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: NormalMeanPrecision

+	out	μ(in1) :: MvNormalMeanPrecision	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: MvNormalMeanPrecision

+	out	μ(in1) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

+	out	μ(in1) :: MvNormalWeightedMeanPrecision{T1, M, P} where {M<:AbstractVector{T1}, P<:AbstractMatrix{T1}}	Nothing

μ(in2) :: MvNormalWeightedMeanPrecision{T2, M, P} where {M<:AbstractVector{T2}, P<:AbstractMatrix{T2}}

+	out	μ(in1) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(in2) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

+	out	μ(in1) :: Distributions.Distribution	Nothing

μ(in2) :: Distributions.Distribution

+	out	μ(in1) :: MvNormalWeightedMeanPrecision	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(in2) :: PointMass

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: MvNormalWeightedMeanPrecision

+	out	μ(in1) :: PointMass	Nothing

μ(in2) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

+	in2	μ(out) :: Any

μ(in1) :: Any

+	in1	μ(out) :: MvNormalWeightedMeanPrecision{T1, M, P} where {M<:AbstractVector{T1}, P<:AbstractMatrix{T1}}	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: MvNormalWeightedMeanPrecision{T1, M, P} where {M<:AbstractVector{T1}, P<:AbstractMatrix{T1}}	Nothing

μ(in2) :: MvNormalWeightedMeanPrecision{T2, M, P} where {M<:AbstractVector{T2}, P<:AbstractMatrix{T2}}

+	in1	μ(out) :: MvNormalMeanPrecision	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: MvNormalMeanPrecision

+	in1	μ(out) :: NormalMeanPrecision	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: NormalMeanPrecision

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: PointMass

+	in1	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Nothing

μ(in2) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

+	in1	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Nothing

μ(in2) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: MvNormalWeightedMeanPrecision

+	in1	μ(out) :: PointMass	Nothing

μ(in2) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T

*	out	μ(A) :: PointMass{<:LinearAlgebra.UniformScaling}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

*	out	μ(A) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:Real}

*	out	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: MvNormalWeightedMeanPrecision

*	out	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: MvNormalMeanPrecision

*	out	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: MvNormalMeanCovariance

*	out	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	out	μ(A) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractVector}

*	out	μ(A) :: PointMass{<:AbstractVector}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	out	μ(A) :: F	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractMatrix}

*	out	μ(A) :: PointMass{<:AbstractMatrix}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: F

*	out	μ(A) :: GammaDistributionsFamily	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:Real}

*	out	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: GammaDistributionsFamily

*	out	μ(A) :: PointMass	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass

*	out	μ(A) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	out	μ(A) :: Distributions.UnivariateDistribution	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Distributions.UnivariateDistribution

*	in	μ(out) :: PointMass	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass

*	in	μ(out) :: GammaDistributionsFamily	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:Real}

*	in	μ(out) :: F	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:Real}

*	in	μ(out) :: MvNormalMeanCovariance	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:AbstractMatrix}

*	in	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:AbstractMatrix}

*	in	μ(out) :: MvNormalMeanCovariance	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:AbstractVector}

*	in	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: PointMass{<:AbstractVector}

*	in	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	in	μ(out) :: Distributions.UnivariateDistribution	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(A) :: Distributions.UnivariateDistribution

*	in	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	in	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: MvNormalMeanCovariance

*	in	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: MvNormalMeanPrecision

*	in	μ(A) :: PointMass{<:Real}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: MvNormalWeightedMeanPrecision

*	in	μ(A) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: PointMass{<:Real}

*	in	μ(A) :: PointMass{<:LinearAlgebra.UniformScaling}	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(out) :: Union{NormalDistributionsFamily{T}, GaussianDistributionsFamily{T}} where T

*	A	μ(out) :: PointMass	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass

*	A	μ(out) :: GammaDistributionsFamily	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:Real}

*	A	μ(out) :: F	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:Real}

*	A	μ(out) :: MvNormalMeanCovariance	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractMatrix}

*	A	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractMatrix}

*	A	μ(out) :: MvNormalMeanCovariance	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractVector}

*	A	μ(out) :: Union{MultivariateGaussianDistributionsFamily{T}, MultivariateNormalDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: PointMass{<:AbstractVector}

*	A	μ(out) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Union{UnivariateNormalDistributionsFamily{T}, UnivariateGaussianDistributionsFamily{T}} where T

*	A	μ(out) :: Distributions.UnivariateDistribution	Union{Nothing, MatrixCorrectionTools.AbstractCorrectionStrategy}

μ(in) :: Distributions.UnivariateDistribution